about halfway through an 'any-utf' implementation
it uses an index caching system that I am far, far too proud of
to provide iteration-like performance through the non-iterating
char seq interface when it is being used to iterate chars.

it also works well for ascii-only strings, and should be pretty
okay for all but the worst possible cases -- which is pretty in
line with my degrading performance correlated with rarer cases
goal.