add benchmarks sub-module (not as multi-module)
adds the 'bench' module for running jmh benchmarks, jol analysis,
and other performance toys. It is in the bench directory, and has
a dependency on basis, but this commit does not change basis into
a multi-module project. This means that the bench module must be
compiled separately, and may be a little confusing, but this was
the simplest and most non-destructive method we found.

this setup may be improved upon in the future. In the meantime, the
pom file for the bench module is intentionally minimalist and
template-like for easy re-use in adding similar modules to other
projects.

a single benchmark is included mostly as an example. It is a
cleaned up version of some tests I ran on a basis method from my
local java-test project. It has some interesting results, but not
terribly important ones.