don't swallow IOExceptions in IOWrap
previously it would swallow errors (using ex.printStackTrace) and
return the stream it was given without comment. The only possible
use for this I can imagine is that if everything else was fine
with your stream, but it wasn't actually compressed, it would
handle that for you. That makes a lot of assumptions about what
kind of IOExceptions can occur though -- ie. you wouldn't want to
swallow any originating from the source stream. This behavior
could be mostly preserved by swallowing only ZipExceptions, but
if that was the intention, then why would the output stream
wrapper also swallow io exceptions? Since there is no defined
behavior via comments, I am changing it to the most sensible
behavior possible without changing the method signature.