Fix Varint#writeSignedVarLong(long value, ByteBuf out)
Looks like a copy/paste error occured here. I've added unit tests
to verify the correct behavior. Initially I was concerned that we
were going to be changing the underlying representation of
signedVarLong() for anyone already using it. But the unit tests
confirmed that encoding had the bug and decoding did not have the bug.
Which means this method does not work at all in its previous state.