add stream util class w/ iterator -> stream helper
for some reason the jdk seems to be (?) missing a few helper methods
to bridge this gap. The iterator version seems particularly out of
the way and overly verbose. Maybe they didn't want people to use these
instead of (probably more efficient) Collections.stream and so on when
they are available, but doing this makes much more sense in many of our
use cases.