add "Scaling" annotation
The intention is to provide more information about the performance
effects of using various methods and objects (both those in libraries
and those in large projects).

This is often tied closely to how much optimization was done on a piece
of code, and a similar annotation may have also mostly sufficed (and may
still be forthcoming). However, sometimes even optimized code is
deliberately geared to wards one of these use cases -- either in clarity
or in function.