Disk queue option to purge on init errors.
This is for applications that are using the the disk backed queue
for off-heap storage and don't care about the durability of the queue.
With this option enabled then any initialization errors will wipe
the contents of the external directory. In general this is a bad idea,
unless your application doesn't care, in which case it's ok. Ref T59413